chapter_004/src/test/java/ru/job4j/lambda/attachmentsort/DiapasonTest.java
package ru.job4j.lambda.attachmentsort;

import org.junit.Test;

import java.util.Arrays;
import java.util.List;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

/**
 * Test.
 *
 * @author fourbarman (mailto:maks.java@yandex.ru)
 * @version 1
 * @since 13.04.2020
 */
public class DiapasonTest {
    /**
     * Test when Linear function.
     */
    @Test
    public void whenLinearFunctionThenLinearResults() {
        List<Double> result = Diapason.diapason(5, 8, x -> 2 * x + 1);
        List<Double> expected = Arrays.asList(11D, 13D, 15D);
        assertThat(result, is(expected));
    }

    /**
     * Test when square function.
     */
    @Test
    public void whenSquareFunctionThenSquareResults() {
        List<Double> result = Diapason.diapason(5, 8, x -> x * x + 1);
        List<Double> expected = Arrays.asList(26D, 37D, 50D);
        assertThat(result, is(expected));
    }

    /**
     * Test when powering function.
     */
    @Test
    public void whenPowFunctionThenPowResults() {
        List<Double> result = Diapason.diapason(5, 8, x -> Math.pow(2, x) + 1);
        List<Double> expected = Arrays.asList(33D, 65D, 129D);
        assertThat(result, is(expected));
    }
}
